[focs][py] rename species/species_macros/stealth -> .../stealth_trait
eases mypy. is a bit less ambiguous
