+++ default/scripting/species/SP_BANFORO.focs.py
@@ -17,7 +17,7 @@ from species.species_macros.planet_shields import AVERAGE_PLANETARY_SHIELDS
 from species.species_macros.population import AVERAGE_POPULATION, LIGHT_SENSITIVE
 from species.species_macros.research import AVERAGE_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import GOOD_STEALTH
+from species.species_macros.stealth_trait import GOOD_STEALTH
 from species.species_macros.stockpile import AVERAGE_STOCKPILE
 from species.species_macros.supply import AVERAGE_SUPPLY
 from species.species_macros.troops import AVERAGE_DEFENSE_TROOPS
