+++ default/scripting/species/SP_EXPERIMENTOR.focs.py
@@ -15,7 +15,7 @@ from species.species_macros.planet_shields import AVERAGE_PLANETARY_SHIELDS
 from species.species_macros.population import GOOD_POPULATION
 from species.species_macros.research import ULTIMATE_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import ULTIMATE_STEALTH
+from species.species_macros.stealth_trait import ULTIMATE_STEALTH
 from species.species_macros.stockpile import ULTIMATE_STOCKPILE
 from species.species_macros.supply import AVERAGE_SUPPLY
 from species.species_macros.troops import ULTIMATE_DEFENSE_TROOPS
