+++ default/scripting/species/SP_FIFTYSEVEN.focs.py
@@ -26,7 +26,7 @@ from species.species_macros.planet_size import NOT_LARGE_PLANET
 from species.species_macros.population import BAD_POPULATION
 from species.species_macros.research import ULTIMATE_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import GOOD_STEALTH
+from species.species_macros.stealth_trait import GOOD_STEALTH
 from species.species_macros.stockpile import AVERAGE_STOCKPILE
 from species.species_macros.supply import AVERAGE_SUPPLY
 from species.species_macros.troops import BAD_DEFENSE_TROOPS
