+++ default/scripting/species/SP_KILANDOW.focs.py
@@ -34,7 +34,7 @@ from species.species_macros.planet_shields import AVERAGE_PLANETARY_SHIELDS
 from species.species_macros.population import AVERAGE_POPULATION
 from species.species_macros.research import GREAT_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import BAD_STEALTH
+from species.species_macros.stealth_trait import BAD_STEALTH
 from species.species_macros.stockpile import AVERAGE_STOCKPILE
 from species.species_macros.supply import AVERAGE_SUPPLY
 from species.species_macros.troops import AVERAGE_DEFENSE_TROOPS
