+++ default/scripting/species/SP_MISIORLA.focs.py
@@ -18,7 +18,7 @@ from species.species_macros.planet_shields import AVERAGE_PLANETARY_SHIELDS
 from species.species_macros.population import BAD_POPULATION
 from species.species_macros.research import BAD_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import BAD_STEALTH
+from species.species_macros.stealth_trait import BAD_STEALTH
 from species.species_macros.stockpile import BAD_STOCKPILE
 from species.species_macros.supply import BAD_SUPPLY
 from species.species_macros.troops import BAD_DEFENSE_TROOPS
