+++ default/scripting/species/SP_NYMNMN.focs.py
@@ -24,7 +24,7 @@ from species.species_macros.native_fortification import DEFAULT_NATIVE_DEFENSE
 from species.species_macros.population import BAD_POPULATION
 from species.species_macros.research import AVERAGE_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import GOOD_STEALTH
+from species.species_macros.stealth_trait import GOOD_STEALTH
 from species.species_macros.stockpile import AVERAGE_STOCKPILE
 from species.species_macros.supply import BAD_SUPPLY
 from species.species_macros.troops import BAD_DEFENSE_TROOPS
