+++ default/scripting/species/SP_SETINON.focs.py
@@ -15,7 +15,7 @@ from species.species_macros.native_fortification import DEFAULT_NATIVE_DEFENSE
 from species.species_macros.population import GOOD_POPULATION
 from species.species_macros.research import AVERAGE_RESEARCH
 from species.species_macros.shields import STANDARD_SHIP_SHIELDS
-from species.species_macros.stealth import GREAT_STEALTH
+from species.species_macros.stealth_trait import GREAT_STEALTH
 from species.species_macros.stockpile import AVERAGE_STOCKPILE
 from species.species_macros.supply import AVERAGE_SUPPLY
 from species.species_macros.troops import NO_DEFENSE_TROOPS, NO_OFFENSE_TROOPS
