+++ default/scripting/species/species_macros/stealth_trait.py
